refactor(entry): split launcher into thin facade
